Remove unnecessary classes
We now have just one implementation of config requests, so remove
unnecessary abstract classes
